[RTL] Implement exception and unwind code for amd64
Messages from Timo:

[RTL] Implement RtplUnwindInternal and wrap RtlUnwindEx and RtlDispatchException around it

Based on the great description in this blog article: http://www.nynaeve.net/?p=106
A few things are done differently

* Add support for version 2 unwind info
* Implement UnwindOpSlots() and use it (based on https://github.com/dotnet/coreclr/blob/master/src/unwinder/amd64/unwinder_amd64.cpp)
* Fix handling of UWOP_PUSH_MACHFRAME

[RTL] Implement RtlCaptureNonVolatileContextPointers and RtlSetUnwindContext

Instead of only checking for the TargetFrame, also check for a mode change, i.e. RIP went from kernel to user,
in which case the target frame was not reached yet, because it was too large, but processing can't continue here.

[RTL] Use movaps instead of movdqa in RtlCaptureContext

It does the same thing, but is one byte shorter

Co-authored-by: Timo Kreuzer <timo.kreuzer@reactos.org>